requirements.txt
flask_sqlalchemy
flask_login
flask_migrate
pymysql